requirements.txt
awscli==1.29.7
beautifulsoup4==4.8.2
GitPython==3.1.41
boto3==1.28.7
Flask==2.3.3
glean-parser~=14.1.0
google-cloud-bigquery==3.23.1
google-cloud-storage==2.2.1
gsutil==5.28
Jinja2==3.1.3
jsonschema==3.1.1
python-dateutil==2.8.0
PyYAML==6.0.1
requests==2.31.0
requests_cache==0.5.2
requests_file==1.4.3
schema==0.7.1
urllib3==1.26.18
Werkzeug==2.3.8
yamllint
